refactor: move constants with field lengths into the new field-lengths folder.
